kvm: refactoring of propagating host params
